Work

+++ db/migrate/20171201025115_create_reviews.rb
@@ -2,7 +2,7 @@ class CreateReviews < ActiveRecord::Migration[5.1]
   def change
     create_table :reviews do |t|
       t.text    :comment,   null: false
-      t.decimal :valuation, null: false
+      t.integer :valuation, null: false
       t.integer :plan_id,   null: false
       t.integer :user_id,   null: false
